# openapi_dump.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastGPT RAG API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/collections/list": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "List Collections",
        "operationId": "list_collections_collections_list_post",
        "parameters": [
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionListParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_id}": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "Get Collection Detail",
        "operationId": "get_collection_detail_collections__collection_id__get",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/update": {
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "Update Collection",
        "operationId": "update_collection_collections_update_post",
        "parameters": [
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections": {
      "delete": {
        "tags": [
          "collection"
        ],
        "summary": "Delete Collection",
        "operationId": "delete_collection_collections_delete",
        "parameters": [
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionDelete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_id}/data/push": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Push Data",
        "operationId": "push_data_collections__collection_id__data_push_post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushDataParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_id}/data/list": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "List Data",
        "operationId": "list_data_collections__collection_id__data_list_post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataListParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_id}/data/{data_id}": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Data Detail",
        "operationId": "get_data_detail_collections__collection_id__data__data_id__get",
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Id"
            }
          },
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "data"
        ],
        "summary": "Delete Data",
        "operationId": "delete_data_collections__collection_id__data__data_id__delete",
        "parameters": [
          {
            "name": "data_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Id"
            }
          },
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_id}/data/update": {
      "put": {
        "tags": [
          "data"
        ],
        "summary": "Update Data",
        "operationId": "update_data_collections__collection_id__data_update_put",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Device-Name"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CollectionDelete": {
        "properties": {
          "collectionIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collectionids"
          }
        },
        "type": "object",
        "required": [
          "collectionIds"
        ],
        "title": "CollectionDelete"
      },
      "CollectionListParams": {
        "properties": {
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize",
            "default": 10
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "searchText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searchtext",
            "default": ""
          }
        },
        "type": "object",
        "title": "CollectionListParams"
      },
      "CollectionUpdate": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "externalFileId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalfileid"
          },
          "datasetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datasetid"
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentid"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "forbid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forbid"
          },
          "createTime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createtime"
          }
        },
        "type": "object",
        "title": "CollectionUpdate"
      },
      "DataListParams": {
        "properties": {
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize",
            "default": 10
          },
          "searchText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Searchtext",
            "default": ""
          }
        },
        "type": "object",
        "title": "DataListParams"
      },
      "DataUpdate": {
        "properties": {
          "dataId": {
            "type": "string",
            "title": "Dataid"
          },
          "q": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Q"
          },
          "a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "A"
          },
          "indexes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/IndexItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexes"
          }
        },
        "type": "object",
        "required": [
          "dataId"
        ],
        "title": "DataUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IndexItem": {
        "properties": {
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type",
            "default": "custom"
          },
          "dataId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataid"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "IndexItem"
      },
      "PushDataItem": {
        "properties": {
          "q": {
            "type": "string",
            "title": "Q"
          },
          "a": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "A",
            "default": ""
          },
          "indexes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/IndexItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexes",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "q"
        ],
        "title": "PushDataItem"
      },
      "PushDataParams": {
        "properties": {
          "collectionId": {
            "type": "string",
            "title": "Collectionid"
          },
          "trainingType": {
            "type": "string",
            "title": "Trainingtype",
            "default": "chunk"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "default": ""
          },
          "billId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Billid"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PushDataItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "collectionId",
          "data"
        ],
        "title": "PushDataParams"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}